Add  addBook method to the class to add new items
