+++ index.html
@@ -34,10 +34,10 @@
                 <h2 class="text-center secondary-header">Add a new book</h2>
                 <form action="javascript:void(0)" method="post" class="form">
                     <div class="form-control my-3">
-                        <input class="input-form border" type="text" required name="title" id="book-title" aria-label="book-title" placeholder="Title">
+                        <input class="input-form border" type="text" required minlength="3" name="title" id="book-title" aria-label="book-title" placeholder="Title">
                     </div>
                     <div class="form-control my-3">
-                        <input class="input-form border" type="text" required name="author" id="book-Author" aria-label="book-Author" placeholder="Author">
+                        <input class="input-form border" type="text" required minlength="3" name="author" id="book-Author" aria-label="book-Author" placeholder="Author">
                     </div>
                     <div class="form-control my-3">
                         <button type="submit" class="btn border">Add
@@ -48,7 +48,7 @@
             </div>
         </section>
     </main>
-    <script src="./src/js/main.js"></script>
+    <script type="module" src="./src/js/main.js"></script>
 </body>
 
 </html>
