+++ src/js/Books.js
@@ -1,0 +1,20 @@
+export default class Book {
+  constructor(title, author) {
+    this._title = title;
+    this._author = author;
+  }
+
+  get title() {
+    return this._title;
+  }
+  get author() {
+    this._author;
+  }
+
+  deleteBook() {}
+
+  addBook(storedBooks) {
+    storedBooks.push(this);
+    localStorage.setItem('books', JSON.stringify(storedBooks));
+  }
+}
